==> t4r/01-workflow.ipynb <==
import gc
import cupy as cp
import numpy as np 
from pathlib import Path

import cudf
import nvtabular as nvt
from merlin.dag import ColumnSelector
from merlin.schema import Schema, Tags

import dask.dataframe as dd
from dask_cuda import LocalCUDACluster
from dask.distributed import Client

cluster = LocalCUDACluster(
    n_workers=1,                       
    device_memory_limit="18GB",        
    local_directory="/nvme/scratch/",
)
client = Client(cluster)

SESSIONS_MAX_LENGTH = 20 
MINIMUM_SESSION_LENGTH = 2

item_id = ['aid'] >> nvt.ops.TagAsItemID()
event_type =  ['type'] >> nvt.ops.AddMetadata(tags=[Tags.ITEM])
cat_feats = item_id + ['type'] >> nvt.ops.Categorify()

session_ts = ['ts']

session_time = (
    session_ts >>
    nvt.ops.LambdaOp(lambda col: cudf.to_datetime(col, unit='ms')) >>
    nvt.ops.Rename(name='ts_dt')
)

sessiontime_weekday = (
    session_time >>
    nvt.ops.LambdaOp(lambda col: col.dt.weekday) >>
    nvt.ops.Rename(name='ts_dow')
)

def get_cycled_feature_value_sin(col, max_value):
    value_scaled = (col + 0.000001) / max_value
    value_sin = np.sin(2*np.pi*value_scaled)
    return value_sin

def get_cycled_feature_value_cos(col, max_value):
    value_scaled = (col + 0.000001) / max_value
    value_cos = np.cos(2*np.pi*value_scaled)
    return value_cos

weekday_sin = (
    sessiontime_weekday >>
    (lambda col: get_cycled_feature_value_sin(col + 1, 7)) >>
    nvt.ops.Rename(name='t_dow_sin') >>
    nvt.ops.AddMetadata(tags=[Tags.CONTINUOUS])
)
weekday_cos = (
    sessiontime_weekday >>
    (lambda col: get_cycled_feature_value_cos(col + 1, 7)) >>
    nvt.ops.Rename(name='t_dow_cos') >>
    nvt.ops.AddMetadata(tags=[Tags.CONTINUOUS])
)

class ItemRecency(nvt.ops.Operator):
    def transform(self, columns, gdf):
        for column in columns.names:
            col = gdf[column]
            item_first_timestamp = gdf['prod_first_ts']
            delta_days = (col - item_first_timestamp) / (60*60*24)
            gdf[column + "_age_days"] = delta_days * (delta_days >=0)
        return gdf

    def compute_selector(
        self,
        input_schema: Schema,
        selector: ColumnSelector,
        parents_selector: ColumnSelector,
        dependencies_selector: ColumnSelector,
    ) -> ColumnSelector:
        self._validate_matching_cols(input_schema, parents_selector, "computing input selector")
        return parents_selector

    def column_mapping(self, col_selector):
        column_mapping = {}
        for col_name in col_selector.names:
            column_mapping[col_name + "_age_days"] = [col_name]
        return column_mapping

    @property
    def dependencies(self):
        return ["prod_first_ts"]

    @property
    def output_dtype(self):
        return np.float64

recency_feats = ['ts'] >> ItemRecency()
recency_feats_norm = (
    recency_feats >>
    nvt.ops.LogOp() >>
    nvt.ops.Normalize(out_dtype=np.float32) >>
    nvt.ops.Rename(name='product_recency_day_log_norm')
)

time_features = (
    session_time +
    sessiontime_weekday +
    weekday_sin +
    weekday_cos +
    recency_feats_norm
)

groupby_feats = ['ts', 'session'] + cat_feats + time_features

groupby_features = groupby_feats >> nvt.ops.Groupby(
    groupby_cols=['session'],
    sort_cols=['ts'],
    aggs={
        'aid': ['list', 'count'],
        'type': ['list'],
        'ts': ['first'],
        'ts_dt': ['first'],
        't_dow_sin': ['list'],
        't_dow_cos': ['list'],
        'product_recency_day_log_norm': ['list']
    },
   name_sep="-"
)

groupby_features_list = groupby_features['aid-list', 'type-list', 't_dow_sin-list', 't_dow_cos-list', 'product_recency_day_log_norm-list']

groupby_features_trim = groupby_features_list >> nvt.ops.ListSlice(-SESSIONS_MAX_LENGTH, pad=True)

day_index = (
    session_time >> 
    nvt.ops.LambdaOp(lambda col: ((col - col.min()).dt.days  + 1)) >> 
    nvt.ops.Rename(f=lambda col: "day_index") >>
    nvt.ops.AddMetadata(tags=[Tags.CATEGORICAL])
 )

sess_id = groupby_features['session']

selected_features = sess_id + groupby_features['aid-count'] + groupby_features_trim + day_index

filtered_sessions = selected_features >> nvt.ops.Filter(f=lambda df: df["aid-count"] >= MINIMUM_SESSION_LENGTH)

workflow = nvt.Workflow(filtered_sessions, client=client)

data_path = Path.cwd().parent / 'preprocess/data'
temp_path = Path.cwd() / 'data'
if not temp_path.is_dir():
    temp_path.mkdir()

## LB

lb_in = data_path / 'lb'
lb_out = temp_path / 'lb'
if not lb_out.is_dir():
    lb_out.mkdir()

train_path = lb_in / 'train_parquet'

data = []
for train_file in sorted(train_path.glob('*.parquet')):
    data.append(dd.read_parquet(train_file.as_posix()))

data = dd.concat(data)

item_first_interaction_df = data.groupby('aid').agg({'ts': 'min'}) \
    .reset_index().rename(columns={'ts': 'prod_first_ts'})
data = data.merge(item_first_interaction_df, on=['aid'], how='left').reset_index(drop=True)

del item_first_interaction_df
gc.collect()

dataset = nvt.Dataset(data, npartitions=4)
dataset.shuffle_by_keys('session')
del data
gc.collect()

%%time
workflow.fit_transform(dataset).to_parquet((lb_out / 'processed_nvt').as_posix())

del dataset
gc.collect()

workflow.output_schema

workflow_path = lb_out / 'workflow_etl'
workflow.save(workflow_path.as_posix())

del workflow
gc.collect()

client.shutdown()
client.close()
